Avoid Int overflow in property size estimation and introduce GQL status for too big properties ()
Attempting to store too big properties could result in integer overflows
when trying to run `format.worstCaseSize(value)` or
`format.bestCaseSize(value)`.

This avoid these errors by using longs for the size estimation all the
way. Subsequently, to improve error messages from `BufferFullException`,
this PR introduces GQL status code `22NBF` to clearly communicate that a
property value is too big to be stores.

+++ community/import-api/src/main/java/org/neo4j/batchimport/api/input/PropertySizeCalculator.java
@@ -24,5 +24,5 @@ import org.neo4j.memory.MemoryTracker;
 import org.neo4j.values.storable.Value;
 
 public interface PropertySizeCalculator {
-    int calculateSize(Value[] values, CursorContext cursorContext, MemoryTracker memoryTracker);
+    long calculateSize(Value[] values, CursorContext cursorContext, MemoryTracker memoryTracker);
 }
